Arbitrary views of the password database. New search UI.
Complete overhaul of viewing folders and database entries, such that
any arbitrary hierarchy of groups (or group-like collections!) and
entries can be created via implementing a trait. UI code is now
completely decoupled from the notion of "groups" and "templates," and
instead operates on "containers" and "entries." A container is
something that holds child containers and/or actual DB entries (i.e.
folder/entry structure).

This lays the groundwork for implementing tags, saved searches, etc.